tool/requirements.txt
selenium
texttable
pyyaml
